Add javascript support library
Provide a javascript library for decoding oneseismic results. The
library itself is mostly the decoder compiled with emscripten, a little
bit of binding code, and some usability features. Please be aware that
the module itself is quite rough and needs a lot of love and testing.
Error paths in particular are not checked exhaustively (if at all).

Using oneseismic through the web browser (or even some other js runtime
like node) will be common, and native javascript decoding makes
client-side rendering fast and feasible.

With the library comes a couple of build scripts - driving the build
with cmake is quite painless, and the emscripten.dockerfile makes it
possible to run even without installing the emscripten sdk directly on
the host.

No testing hooks or testing strategy is included in the build scripts,
nor any deployment. I have yet to figure out how to distribute the
library, only how to build it.

The library can target both webassembly and "pure" js (asm.js), which is
easily controllable from the cmake file.

Like the python code, the javascript library is LGPL licenced - this
includes the decoder itself, even though the decoder is in the core/
tree.