Add javascript CI job, a few tests
Introduce some basic test infrastructure and a github actions job to
make sure the javascript support lib builds and works as expected.

The tests are built for node to make them easier to run in a command
line and CI setting. No framework or library is used; partly due to my
infamiliarity with the js environment, partially because it would take
attention away from the infrastructure of this patch.

Testing that the decoder itself works is probably not that interesting
from the JS perspective, only that the binding code and emscripten works
right under certain conditions - most notably when buffering and
processing small chunks of data.