Add statisical sampledata to the manifest
It turns out that statistical data about the samples, such as the range
of the sample-values within a cube can be useful information for the
end-users. However, end-users would have to scan the entire cube, which
is often more expensive than it's worth. However, recording and
embedding this information into the manifest is almost free for us to
do, as we scan the entire file on upload anyways.

This commit adds two new fields to the manifest; sample-value-min &
sample-value-max.