Formalise (and share) fetch+download queue
Clean up and separate the download queue features of cmd/fetch, and more
obviously share the download queue between requests.

The idea was always to share the download queue between requests;
whenever a download slot (controlled by opts.jobs) was free, the next
task should be pulled from redis. This was awkward when the goroutines
that ultimately lead to the download happening were spawned in the run()
function, making it unclear if scheduling and resource capping actually
worked the way it was supposed to.

Enter fetch [1], which hides a lot of the channel stuff, and provides
the higher-level enqueue() (and startWork(), for convenience). Rather
than run() having to spawn goroutines and trying to respect the size of
the worker pool, tasks are submitted, and read off of a private channel
for the session.

Overall this approach is much nicer, and should be more responsive in
scheduling downloads.

[1] There is probably a much better name for this concept, and the
    second one is conceived I encourage a renaming fiesta.