Cache fragments on worker nodes
As far as distributed systems go, oneseismic is on the easier side since
most data is read-only [1]. The bulk of the work in a request is to
download fragments from the blob store, but since fragments really
shouldn't change, the fragments can be aggressively cached.

Designing it was a bit back and forth. I started prototyping it with a
localhost redis instance. This proved a significant speedup, but relies
on a running external redis and a bunch of copying over loopback. Redis
is great because it implements all kinds of expiration and eviction
algorithms, but comes at a deployment and runtime cost.

The next step was moving the cache in-process. This means fresh or
restarted nodes start out with an empty cache (e.g. during auto
scaling), and that's fine for now. Two caching libraries looked viable:

* https://github.com/patrickmn/go-cache
* https://github.com/dgraph-io/ristretto

I prototyped with both, but ristretto won out since it has a boatload of
users, implements bounded size, staged add, and least-frequently-used
eviction policies, which is a better fit for oneseismic than pure
expiration-based eviction.

The job scheduling does not try to give jobs to nodes that have
fragments in cache. This could be an improvement in the future, but is
difficult to get right and requires a lot of inter-node communication.
For now, the scheme is for nodes to cache whatever passes through it.
This should work reasonably well in practice.

Caching fragments means 5-10x end-to-end speedup when the (distributed)
cache is reasonably hot. This makes common queries, and especially
interesting section of interesting volumes, a lot faster.

[1] unless the system is somehow unhealthy, fragments should not change