Use nginx for routing in docker-compose
Put an nginx instance in front also in docker-compose, rather than
connecting to the API-node directly. This is not a functional change at
this stage, but paves the way for moving the /result endpoint into its
own process, which will need routing if the api nodes and the result
are to be served from the same domain.

This complicates the test setup slightly, but brings it closer to the
production environment, which is probably a good thing.