Fix tests failing for unintended reason
Assumption is that tests were meant to assure minimum required setup,
but were not updated due to missing "positive" setup or message check.

It's challenging to come up with a maintainable solution fit for testing
purpose as query keys might get updated often, might be easy to miss and
json is tiresome to update. Plain tests unfortunately have all these
disadvantages.

The best option seemed to be creation of one query for all relevant
tests and modifying it in place with json library whenever needed.

Additionally, there is a question on the form on such tests.
1. Current way of writing tests "for" + "Section" is not efficient:
"for" and "Section" do not play together well, resulting in n² time
instead of n. As we do not have many test cases, it shouldn't matter
much though.
2. Using Generator instead of for loop helps, but due to nature of
Generator it in some cases requires assurance that all underlying data
lives between different iterations of generator. Recommended catch2 way
of dealing with that is not sufficient in all cases, and it is not clear
if existing workarounds are sustainable.
3. Last option is just removing "Section" - aka reporting and running
all the tests together with info messages. This approach was chosen in
one case due to its simplicity.