Disallow packing query from within c++ code
At the moment query object serves to pass information from go to c++.
Object is created (packed) in go and is unpacked in c++. Thus packing
operation is never supposed to happen in c++ and only the basic_query
version of it was implemented.

As this operation is not supposed to exist by design, code is updated
respectively.