Assert 3-dimensional cube and fragments shape
While during query processing we assume that downloaded manifest is
correct and contains at least one valid fragments data entry (which at
the moment is supposed to be used), one-seismic doesn't have to be 3D.
It's not unlikely that other dimensionality will be supported in future.

For a moment just assert on it and keep in mind that solution might
change if bigger support is given to non 3D-cubes.