Move first trace into scan-loop
The scan function read the first trace before going into it's regular
trace-loop. This led to a bug when scanning the sample values, as the
first trace where not included in the statistic.

The first trace is moved into the loop such that all traces are
proccessed the same way.