Allow passing multiple scanners to scan
Although this change do make the scan function a little more versatile,
the main reasoning for it is to maintain code-quality as the number of
scan operations increase. It allows better separation of distinct scan
operations, such as line number scanning. Extending scanners with
logging, defaults and options is also possible without cluttering the
general interface and logic too much. These are all features that will
come in handy if we want to make the scan/upload process more
interactive by allowing users to intercept and change the scan output
before proceeding with upload.