Make scanner a pure interface class
With this commit we change the way we implement scanners. The scanner
class is repurposed as a pure no-op class that scanners derive their
interface from. All scan logic is moved into separate scanners. In this
new setup each scanner implements a specific scan-operation and is
independent of other scanners. The main purpose is code-clarity and
making it easy to add new scanners and extending them with e.g. logging,
defaults or options easier without to much clutter.