Run local integration tests on CI
Run tests as a separate CI job.
Unfortunately there are issues with just extending the build job due to
not-so-easy docker handling on Windows and MacOS. As of now, client code
must be rebuild again. Attempt to reduce code repetition may be done in
the future.

Additionally commit brings back go tests, which were not executed on CI
for some time, as part of the server build.