Update setup for version tests (client)
It has been agreed that as we do not want to pin requirements.dev
versions as for now, we need to use other strategy with
regards to client version testing.

Starting from the next release, releases will be tagged in source
control. Python wheel with requested tag will be downloaded through pip
and code with the same tag will be checked out in source control, to
obtain tests relevant for the version.