Initial commit for the oneseismic catalogue API
This commit is huge, but there are really no good way of splitting it up
into atomic commits that makes sense.

This commit introduces a new API endpoint for oneseismic, the Catalogue
API. It's main motivation is cube-discoverability. It lets the caller
search the oneseismic manifests to find cubes of interest. To achieve
this, this GraphQL API provides two ways to filter manifests:

  1) JSON filtering. Filter the manifest based on values in the JSON
     fields. Any manifest field can be used for filtering results.

  2) Geometric filtering. Find all manifests that intersects with a
     Point, Linestring or Polygon defined by the caller.

The catalogue architecture
--------------------------

While this commit only concern itself with the GraphQL API, some context
to the overall architecture is useful. There are 3 main parts to it:

  1) Ingestion: The ingestion service reacts to blob events such as
     creation, deletion and changes to existing blobs (manifest.json)
     and updates the catalogue storage. In effect it maintains a mirror
     of all the manifests in Blob Store in a Postgres DB.

     What follows naturally from that, is the ability to route manifests
     from different Storage Accounts to different postgres DB's. I.e.
     the manifests from one Storage Account can be mirrored in their own
     isolated DB, or in a larger DB together with manifests from other
     Storage Account. Creating one large search index.

  2) Storage: The manifests are stored as JSONB objects in Postgres
     DB's. This is super-convenient as there is no need for a database
     schema for the manifest fields. Additionally, we create a PostGIS
     polygon of the cube, which gives us out-of-the box support for
     geometric queries.

  3) GraphQL API: Postgres provides all the querying we need. The API is
     basically just a translation layer, providing a nice GraphQL API,
     and translating requests into SQL before querying the DB. It also
     handles auth.

The core of oneseismic have no dependencies to the catalogue. The
catalogue is served from it's own binary. That means that the catalogue
can be regarded as an optional add-on to oneseismic, for applications
wanting better discoverability of their data.

Authentication and Authorization
--------------------------------

Unfortunately, the SAS tokens used by the main query API is a poor fit
for a catalogue like this. The manifests are all stored in different
Azure Blob Store Containers, meaning one would require a Storage
Account wide SAS token in order have access to more than one manifest.
We could restrict the access again by using user delegation SAS, but it
has it's problems too.

We therefore opt for an oauth2 flow, using access tokens and app
registrations. Although its rather annoying to go for a different auth
solution in this API, it seams to be the better option. This can of
course be revised in the future.