Add a basic azblob implementaion of StorageClient
This commit introduces an azure blobstore specific implementation of the
storage facade StorageClient. It's very simple at the moment, with no
support for caching and only support credentials in the form of
query-parameters. Once it's a bit more mature the rest of the system
(query api and fetch) will adopt and use this implementation.