Add optional caching to AzStorage
This commit copies the caching implementation from pipeline.go and
re-implement it's use inside AzStorage.Get.

The copying of the cache-implementation should be regarded as a 2-step
move. For now it's simply copied to avoid any hazzle with unexported
symbols across multiple go packages. But once fetch/pipeline adopts the
use of the storage facade, it will be removed from there.